helpers: format vagrant-cloud.pl uploader properly
also use perl variable interpolation between double quotes to construct
rest endpoints URIs

A note about variable interpolation:
since the delimiter between the variables ( / or . ) is not a valid
character for a variable name, the interpreter is able to grok the
begin and end of the interpolated variable without using curly braces.
This is similar to shell where
echo ".$HOME./$USER/"
does the $HOME and $USER variables do not need to be protected with curly
braces.